+++ cmd/template/nodepool/runner.go
@@ -2,7 +2,6 @@ package nodepool
 
 import (
 	"context"
-	"fmt"
 	"io"
 	"os"
 	"strings"
@@ -123,7 +122,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		config.ReleaseComponents = releaseCollection.ReleaseComponents(r.flag.Release)
 
 		if r.flag.Provider == key.ProviderAzure {
-			config.Namespace = fmt.Sprintf("org-%s", config.Owner)
+			config.Namespace = key.OrganizationNamespaceFromName(config.Owner)
 		}
 	}