+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 - Start publishing a container image of kubectl-gs as giantswarm/kubectl-gs
 
 ### Changed
+
+- Normalize organization name when used as a namespace, to match company requirements.
 - Allow using inactive release versions for templating clusters. This is especially useful for testing cluster upgrades.
 
 ## [0.7.2] - 2020-10-12